Acceso por rol y login, logica de bloqueo de funciones dependiendo del usuario

+++ SistemaIMC/SistemaIMC/Controllers/T_CursoController.cs
@@ -1,64 +1,47 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SistemaIMC.Data;
 using SistemaIMC.Models;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace SistemaIMC.Controllers
 {
+    // Director puede ver
+    [Authorize(Roles = "Administrador del Sistema, Supervisor / Director de la Entidad")]
     public class T_CursoController : Controller
     {
         private readonly TdDbContext _context;
 
-        public T_CursoController(TdDbContext context)
-        {
-            _context = context;
-        }
+        public T_CursoController(TdDbContext context) { _context = context; }
 
-        // GET: T_Curso
         public async Task<IActionResult> Index()
         {
-            var curso = await _context.T_Curso
-                                .Include(c => c.Establecimiento)
-                                .ToListAsync();
-
+            var curso = await _context.T_Curso.Include(c => c.Establecimiento).ToListAsync();
             return View(curso);
         }
 
-        // GET: T_Curso/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var t_Curso = await _context.T_Curso
-                .FirstOrDefaultAsync(m => m.ID_Curso == id);
-            if (t_Curso == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var t_Curso = await _context.T_Curso.FirstOrDefaultAsync(m => m.ID_Curso == id);
+            if (t_Curso == null) return NotFound();
             return View(t_Curso);
         }
 
-        // GET: T_Curso/Create
+        // CREATE - SOLO ADMIN
+        [Authorize(Roles = "Administrador del Sistema")]
         public async Task<IActionResult> Create()
         {
             await LoadEstablecimientosViewBag();
             return View();
         }
 
-        // POST: T_Curso/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Administrador del Sistema")]
         public async Task<IActionResult> Create([Bind("ID_Curso,NombreCurso,ID_Establecimiento")] T_Curso t_Curso)
         {
             if (ModelState.IsValid)
@@ -67,39 +50,26 @@ namespace SistemaIMC.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-
             await LoadEstablecimientosViewBag(t_Curso.ID_Establecimiento);
             return View(t_Curso);
         }
 
-        // GET: T_Curso/Edit/5
+        // EDIT - SOLO ADMIN
+        [Authorize(Roles = "Administrador del Sistema")]
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var t_Curso = await _context.T_Curso.FindAsync(id);
-            if (t_Curso == null)
-            {
-                return NotFound();
-            }
+            if (t_Curso == null) return NotFound();
             return View(t_Curso);
         }
 
-        // POST: T_Curso/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Administrador del Sistema")]
         public async Task<IActionResult> Edit(int id, [Bind("ID_Curso,NombreCurso,ID_Establecimiento")] T_Curso t_Curso)
         {
-            if (id != t_Curso.ID_Curso)
-            {
-                return NotFound();
-            }
-
+            if (id != t_Curso.ID_Curso) return NotFound();
             if (ModelState.IsValid)
             {
                 try
@@ -109,68 +79,39 @@ namespace SistemaIMC.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!T_CursoExists(t_Curso.ID_Curso))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!T_CursoExists(t_Curso.ID_Curso)) return NotFound(); else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
-
             await LoadEstablecimientosViewBag(t_Curso.ID_Establecimiento);
             return View(t_Curso);
         }
 
-        // GET: T_Curso/Delete/5
+        // DELETE - SOLO ADMIN
+        [Authorize(Roles = "Administrador del Sistema")]
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var t_Curso = await _context.T_Curso
-                .FirstOrDefaultAsync(m => m.ID_Curso == id);
-            if (t_Curso == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var t_Curso = await _context.T_Curso.FirstOrDefaultAsync(m => m.ID_Curso == id);
+            if (t_Curso == null) return NotFound();
             return View(t_Curso);
         }
 
-        // POST: T_Curso/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Administrador del Sistema")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var t_Curso = await _context.T_Curso.FindAsync(id);
-            if (t_Curso != null)
-            {
-                _context.T_Curso.Remove(t_Curso);
-            }
-
+            if (t_Curso != null) _context.T_Curso.Remove(t_Curso);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private async Task LoadEstablecimientosViewBag(object selectedEstablecimiento = null)
         {
-            // Se usa ID_Establecimiento como nombre del ViewBag para que sea intuitivo
-            ViewBag.ID_Establecimiento = new SelectList(
-                await _context.T_Establecimientos.OrderBy(e => e.NombreEstablecimiento).ToListAsync(),
-                "ID_Establecimiento", // Valor de la opción
-                "NombreEstablecimiento", // Texto de la opción
-                selectedEstablecimiento // Opción seleccionada
-            );
+            ViewBag.ID_Establecimiento = new SelectList(await _context.T_Establecimientos.OrderBy(e => e.NombreEstablecimiento).ToListAsync(), "ID_Establecimiento", "NombreEstablecimiento", selectedEstablecimiento);
         }
-        private bool T_CursoExists(int id)
-        {
-            return _context.T_Curso.Any(e => e.ID_Curso == id);
-        }
+        private bool T_CursoExists(int id) => _context.T_Curso.Any(e => e.ID_Curso == id);
     }
 }
